feat: 상대방 일정 삭제 불가능

+++ src/components/schedule-calendar/store/ScheduleCalendarStore.ts
@@ -5,10 +5,12 @@ import { DateUtil } from '@utils/DateUtil';
 import { ScheduleCalendarRequest } from '@requests/ScheduleCalendarRequest';
 import { toast } from '@components/common/toast/Toast';
 import { dialog } from '@components/common/dialog/Dialog';
+import UserStore from '@stores/UserStore';
 
 @autobind
 export default class ScheduleCalendarStore {
     private static _instance: ScheduleCalendarStore;
+    private _userStore = UserStore.instance;
 
     @observable private readonly _todayCalendarDate: CalendarDate;
 
@@ -212,6 +214,11 @@ export default class ScheduleCalendarStore {
 
     @action
     deleteSchedule(schedule: Schedule) {
+        if (!this._userStore.isMe(schedule.owner)) {
+            toast.show('상대방의 일정은 삭제할 수 없어요.', 'error');
+            return;
+        }
+
         dialog.confirm('일정을 삭제하시겠어요?', () => {
             (async () => {
                 await ScheduleCalendarRequest.deleteSchedule(schedule);
